requirements.txt
tk
ttkbootstrap
Pillow
requests
pandas
psutil
python-docx
graphviz
google-generativeai
pywin32